#### PR description:
This PR introduces dynamic memory caps for the LS, T3, and T5 objects, reducing LST's memory footprint and eliminating truncation of LST track primitives due to underallocation.

### Memory Reduction (0.8 GeV, 10 event average)
The numbers below are the average sums over 10 events of the memory caps for all modules in an event. Previously, most space that was allocated was not used to avoid possible truncation (eg, only O(6k) T5's are created per event, but 500k elements were allocated in the T5 arrays using the constant caps). This PR fixes this approach by first doing quick counts using a subset of the cuts in the actual creation kernels to get a better estimate of how much space needs to be allocated per module.

First number is using only the static caps (ie prior to https://github.com/cms-sw/cmssw/pull/47232), second is current, third is this PR.

T5's - 502,000 -> 120,000 -> 57,000 (89% decrease)
T3's - 1,980,000 -> 737,000 -> 409,000 (79% decrease)
LS's - 1,931,000 -> 858,000 -> 532,000 (72% decrease)

### Truncation Reduction (0.8 GeV, 10 event sum)

First number is current, second is this PR.

T5's - 0 -> 0 (current limits are already set very high)
T3's - 1,349 -> 0
LS's - 16,488 -> 0

Timing increases on GPU by ~14% (3.6 ms -> 4.1 ms) for single stream and ~6% (1.7 ms -> 1.8 ms) for 8 streams. Most of the timing increase on GPU is from reduced truncation (i.e. more objects being produced by LST, causing more computation downstream) rather than overhead from the new kernels that calculate dynamic caps for memory allocation. Timing with CPU backend increases by a similar 6-7%.

c.c. @slava77