@cmsbuild 2025-08-07
cms-bot internal usage<!-- bot cache: {"commits":{"91ec7388cb6bbec76902d22e26ab1f1c121ff9fe":{"files":["RecoTracker/LSTCore/interface/MiniDoubletsSoA.h","RecoTracker/LSTCore/interface/SegmentsSoA.h","RecoTracker/LSTCore/interface/TripletsSoA.h","RecoTracker/LSTCore/src/alpaka/LSTEvent.dev.cc","RecoTracker/LSTCore/src/alpaka/Quintuplet.h","RecoTracker/LSTCore/src/alpaka/Segment.h","RecoTracker/LSTCore/src/alpaka/Triplet.h"],"squashed":false,"time":1754507934}},"emoji":{},"last_seen_sha":"91ec7388cb6bbec76902d22e26ab1f1c121ff9fe","signatures":{}} -->