@GNiendorf 2025-08-07
Looks like this will lead to conflicts with https://github.com/cms-sw/cmssw/pull/48377 which changes the inputs to make_device_view.