@fwyzard 2025-08-08
@slava77, if the topic of the next development is different, I would suggest to open a second PR adding one or more commits on top of this one.

Also, if we should review these PRs, please assign them to @cms-sw/heterogeneous-l2 .