@cmsbuild 2025-08-08
-1

**Failed Tests**: RelVals-INPUT 
**Size**: This PR adds an extra 16KB to repository
**Summary**: https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-2edc36/47620/summary.html
**COMMIT**: 91ec7388cb6bbec76902d22e26ab1f1c121ff9fe
**CMSSW**: CMSSW_15_1_X_2025-08-08-1100/el8_amd64_gcc12
**Additional Tests**: CUDA,ROCM
**User test area**: For local testing, you can use `/cvmfs/cms-ci.cern.ch/week1/cms-sw/cmssw/48698/47620/install.sh` to create a dev area with all the needed externals and cmssw changes.


## RelVals-INPUT

- [136.897](https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-2edc36/47620/runTheMatrixINPUT-results/136.897_RunCosmics2021CRUZET)```136.897_RunCosmics2021CRUZET/step2_RunCosmics2021CRUZET.log```


## Comparison Summary



**Summary**:
 - You potentially [removed 3 lines](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisons/CMSSW_15_1_X_2025-08-08-1100+2edc36/69774/validateJR/logRootQA-events.log) from the logs
 - [Reco comparison results: 0 differences found in the comparisons](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisons/CMSSW_15_1_X_2025-08-08-1100+2edc36/69774/validateJR.html)
 - DQMHistoTests: Total files compared: 57
 - DQMHistoTests: Total histograms compared: 4457190
 - DQMHistoTests: [Total failures: 10199](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisons/CMSSW_15_1_X_2025-08-08-1100+2edc36/69774/validateJR.html)
 - DQMHistoTests: Total nulls: 0
 - DQMHistoTests: Total successes: 4446971
 - DQMHistoTests: Total skipped: 20
 - DQMHistoTests: Total Missing objects: 0
 - DQMHistoSizes: Histogram memory added: 0.0 KiB( 56 files compared)
 - Checked 240 log files, 206 edm output root files, 57 DQM output files
 - [TriggerResults](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisons/CMSSW_15_1_X_2025-08-08-1100+2edc36/69774/triggerResults): no differences found


## CUDA Comparison Summary



**Summary**:
 - You potentially [removed 1 lines](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisonsCUDA/CMSSW_15_1_X_2025-08-08-1100+2edc36/69773/validateJR/logRootQA-events.log) from the logs
 - [Reco comparison results: 0 differences found in the comparisons](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisonsCUDA/CMSSW_15_1_X_2025-08-08-1100+2edc36/69773/validateJR.html)
 - DQMHistoTests: Total files compared: 9
 - DQMHistoTests: Total histograms compared: 117858
 - DQMHistoTests: [Total failures: 19447](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisonsCUDA/CMSSW_15_1_X_2025-08-08-1100+2edc36/69773/validateJR.html)
 - DQMHistoTests: Total nulls: 0
 - DQMHistoTests: Total successes: 98411
 - DQMHistoTests: Total skipped: 0
 - DQMHistoTests: Total Missing objects: 0
 - DQMHistoSizes: Histogram memory added: 0.0 KiB( 8 files compared)
 - Checked 32 log files, 36 edm output root files, 9 DQM output files
 - [TriggerResults](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisonsCUDA/CMSSW_15_1_X_2025-08-08-1100+2edc36/69773/triggerResults): no differences found


## ROCM Comparison Summary



**Summary**:
 - No significant changes to the logs found
 - [Reco comparison results: 0 differences found in the comparisons](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisonsROCM/CMSSW_15_1_X_2025-08-08-1100+2edc36/69772/validateJR.html)
 - DQMHistoTests: Total files compared: 9
 - DQMHistoTests: Total histograms compared: 117858
 - DQMHistoTests: [Total failures: 20862](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisonsROCM/CMSSW_15_1_X_2025-08-08-1100+2edc36/69772/validateJR.html)
 - DQMHistoTests: Total nulls: 0
 - DQMHistoTests: Total successes: 96996
 - DQMHistoTests: Total skipped: 0
 - DQMHistoTests: Total Missing objects: 0
 - DQMHistoSizes: Histogram memory added: 0.0 KiB( 8 files compared)
 - Checked 32 log files, 36 edm output root files, 9 DQM output files
 - [TriggerResults](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisonsROCM/CMSSW_15_1_X_2025-08-08-1100+2edc36/69772/triggerResults): no differences found