@jfernan2 2025-08-08
assign heterogeneous